refactor: request요청, 복수 api 다루기 위한 클래스들..
